Refactored interpreters to make them easier to load. Updated docs with more information about interpreters.
change-type: patch
