+++ interpreters/cli.js
